create seprate components
